code/Eel_GUI/requirements.txt
altgraph==0.17.2
beautifulsoup4==4.9.3
bottle==0.12.19
bottle-websocket==0.2.9
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
dictdiffer==0.9.0
dlib==19.22.1
Eel==0.14.0
face-recognition==1.3.0
face-recognition-models==0.3.0
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
google-api-core==2.0.1
google-auth==2.0.2
google-cloud-texttospeech==2.5.2
google-cloud-vision==2.4.4
googleapis-common-protos==1.53.0
greenlet==1.1.1
grpcio==1.40.0
gTTS==2.2.3
idna==3.2
joblib==1.0.1
mysql-connector-python==8.0.26
nltk==3.6.5
numpy==1.21.2
opencv-python==4.5.3.56
packaging==21.0
Pillow==8.3.2
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
pydub==0.25.1
pyenchant==3.2.1
pyparsing==2.4.7
python-slugify==5.0.2
regex==2021.8.28
requests==2.26.0
rsa==4.7.2
selenium==3.141.0
simpleaudio==1.0.4
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
text-unidecode==1.3
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.62.2
typing-extensions==3.10.0.1
tzlocal==3.0
urllib3==1.26.6
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
